rename info to app_info & add extra routes
